Fix MarkupSimplifier
The RemoveElementsForDocumentComparison method tries to change the root
elements of the CoreFilePropertiesPart and ExtendedFilePropertiesPart,
both of which are optional in a Word document (although Microsoft Word
always creates them). In case a WordprocessingDocument does not have
any of these parts, a NullReferenceException is thrown within the
PtOpenXmlUtil.GetXDocument method. That method also just catches an
exception and directly rethrows it.

This commit fixes the MarkupSimplifier bug and simplifies the
GetXDocument method by removing the unnecessary try/catch clause.

Fixes #122