Enhance PtOpenXmlExtensions
When using the Open XML PowerTools in conjunction with the strongly
typed classes provided by the Open XML SDK (e.g., Document, Paragraph,
or Run), users might run into unexpected issues.

Firstly, there is an issue when using the strongly typed classes after
having used the PowerTools, e.g., the MarkupSimplifier. When the
PutXDocument method and its overloads write content to an OpenXmlPart's
stream, the user will have to reload the OpenXmlPart's root element to
continue working with the strongly typed classes. However, the user
might not know which OpenXmlParts were changed by the PowerTools, e.g.,
the MarkupSimplifier, which changes multiple parts depending on the
SimplifyMarkupSettings.

Secondly, there is an issue when using the PowerTools and the strongly
typed classes in an alternating fashion. Invoking the GetXDocument
method caches the XDocument read from the OpenXmlPart. After changing
the OpenXmlPart's content through the strongly typed classes or any
other means, invoking the GetXDocument again will not yield the updated
content but the cached XDocument.

The main issue is the caching of data by and resulting statefulness of
the GetXDocument and PutXDocument(OpenXmlPart) methods. Many utility
methods depend on that behaviour, which is thus not easy to change.

This commit adds properties and methods and amends existing methods to
fix the issues mentioned above. It also adds unit tests that
demonstrate the issues and the solutions.

Fixes #129